requirements.txt
pyserial
scipy
openpyxl
